requirements.txt
GitPython
pandas
tabulate
weasyprint==52.5
markdown-it-py